criei novamente o app por um erro que deu no json
