+++ modulo3/labeddit1/projeto-labeddit/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "labeddit",
+  "name": "projeto-labeddit",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
