+++ modulo3/labeddit1/projeto-labeddit/public/manifest.json
