+++ modulo3/labeddit1/projeto-labeddit/src/App.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+function App() {
+  return (
+    <div>
+      eae
+    </div>
+  );
+}
+
+export default App;
